# db/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package dbsqlc

import (
	"time"

	"github.com/google/uuid"
)

type Bookmark struct {
	ID          uuid.UUID  `json:"id"`
	UserID      uuid.UUID  `json:"user_id"`
	Url         string     `json:"url"`
	Title       string     `json:"title"`
	Description *string    `json:"description"`
	Favicon     *string    `json:"favicon"`
	ImageUrl    *string    `json:"image_url"`
	FolderID    *uuid.UUID `json:"folder_id"`
	IsFavorite  bool       `json:"is_favorite"`
	CreatedAt   time.Time  `json:"created_at"`
}

type BookmarkTag struct {
	BookmarkID uuid.UUID `json:"bookmark_id"`
	TagID      uuid.UUID `json:"tag_id"`
}

type Folder struct {
	ID     uuid.UUID `json:"id"`
	UserID uuid.UUID `json:"user_id"`
	Name   string    `json:"name"`
}

type Tag struct {
	ID     uuid.UUID `json:"id"`
	UserID uuid.UUID `json:"user_id"`
	Name   string    `json:"name"`
}

type User struct {
	ID           uuid.UUID `json:"id"`
	Email        string    `json:"email"`
	PasswordHash string    `json:"password_hash"`
	CreatedAt    time.Time `json:"created_at"`
}